Function calls
All function definitions are extracted after parsing for the compiling stage

Type checking for the return value and arity checking (although there
is still no support for arguments).

Function calls can be both statements and expressions, so it is now
possible to have a function that compiles without a return (and will likely
cause problems if so). This behavior will need to be rethought.

Also, still need a way to have true blocks for chained inner statements