File inclusions
File inclusion is done with a similar syntax to other top-level statements:
```
(include "file.snm")
```
However on the backend it is implemented at the time that the file is being
read (essentially before the parsing) making it function like a C preprocessor
directive.

The include just works like a textual substitution of that file. Then, the
combined big file is parsed and compiled.

There is no check for circular inclusions or other reinclusions of a
file.