Long type
Long type partially implemented. Turns out there is a big bug with include,
which treats the bound identifier as a variable (when it is actually a
constant). This will necessitate a later fix.

Long types should be able to be declared as variables, parameters, and return
types, and can also be passed around. Still, the include is the only way to
actually get a value for a long type (besides the initialized value).

The compiler also has gained access to type information which is necessary now
that types are not all the same length and can't necessarily be treated the
same. This makes things very complicated but it also improves the modularity of
the compiler in certain places. There is a new allocation/deallocation strategy
for large data on the
stack betwene function calls and block scope changes.

Local binds are no longer initialized with 0. should I initialize them? Should
you be required to initialize them? Do we just do the C approach which is very
unsafe?

The fix of the included variables opens the door to real named constants,
and even function pointers!