Increment, decrement, and zero out variable statements.
These statements directly modify a variable (and have ! accordingly) and is not
an expression that results in a value. (Basically, none of that `x++` crap; you
get to do one or the other, although maybe it should be implemented anyway).

This feature is an optimization for variables stored globally. Local variables
may still benefit from this feature, but not as much