Removal of the boolean type.
Sorry, but it was just way too unoptimal. Instead of booleans being expressions,
they are now a separate "predicate" thing, cutting out the middleman of handling
boolean types.

Next up will be the ternary operator and a special 0 checker, and perhaps
optimizations to predicate integer comparisons in the case that one or more of
them are immediates.