Byte type.
Byte and word types are now classed together as one integral type. When it comes
to computations, byte and words may be used interchangably. That means that for
a function like
```
(func (f byte ([x byte] [y word])) ... )
```
this is a proper call:
```
(local ([x word] [y byte] [z word])
  (set! z (f x y)))

```

All bytes are
implicitly cast to words when computation is being done with them. The fact that
they are bytes only relates to the initial loading of a byte (the upper byte is
set to 0 to cast it to word) and when storing a byte (the upper byte is thrown
away).

So like
```
(func (g byte ([x byte] [y word]))
  (return (+ x y)))
```
x will be cast to a word and added to y as a word, so the result `(+ x y)` is
also a word. But the return type is byte so this result will be cast to byte
anyway. (A super smart compiler could force y to also be a byte and only do the
computation as bytes, but oh well.)

Internally for the type checker, they are both treated as type int. The actual
behavioral differences are in the compiler only. This may lead to some unhelpful
error messages when there is a type mismatch, but oh well.

Long is not considered an integral type, and will eventually be migrated to the
class of pointer type.