Array indexing operations
Can now get the element at some index of an array and also set it. The only
types that arrays support are byte and word (no long or void).

Reminder to myself to actually develop proper error messages and figure out a
proper pipeline to check all the identifiers in a program