Array and function types, dynamic function calls, and dynamic array accesses.
This is a substantial overhaul of the type system, which allows us to represent
new kinds of types for arrays and functions. Previously, the name of a function
and arrays were considered `long`. Now, we can specify types that are arrays and
functions. For example, we can declare an array of arrays on the top level with
```
(array x (array byte) 10)
```
Note that the name of an array is still the pointer to the first element of the
array. Arrays (and functions) are passed by reference, not by value.

Function pointers have types associated with their parameters and return value.
For example, the function declaration
```
(func (foo void ([x byte] [y (array long)])) ... )
```
would have the type
```
(func void (byte (array long)))
```

Both of these pointer types can be used in place of a `long` (since that's also
its internal representation). `long` is still not an integer; it can still be
imagined like a pointer. However, it cannot be cast back down to an array or
function type.

The grammar of all valid types in the language is
```
TYPE := void
      | byte
      | word
      | long
      | (array TYPE)
      | (func TYPE (TYPE*))
```

With this upgraded type system, we can have dynamic array accesses and function
calls. The syntax of a function call is
```
(f arg1 arg2 ...)
```
Previously, `f` needed to be the actual name of the function, known at
compiletime. Now, it is allowed to be an expression evaluating to a function
pointer (and the arguments must match in type/number to that type).

A similar upgrade has been made to arrays:
```
(array-set! a idx val)
(array-get a idx)
```
so `a` can now be an expression evaluating to an array type.

In the case that `f` and `a` are the direct names of functions and arrays
defined on the top level, it is compiled the same way it was previously. If `f`
and `a` are expressions (and thus dynamic), an evaluation must be made before
the jump/access, which is more expensive. Thus, it is faster to just use direct
labels anyway, but overall this allows for more expressibility in some
situations hopefully.

There has also been some cleaning up in the project, especailly when it comes to
extracting top-level declarations. More to follow.